requirements.txt
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
Django==4.0
django-allauth==0.54.0
django-debug-toolbar==4.2.0
django-environ==0.10.0
django-extensions==3.2.3
django-redis==5.3.0
oauthlib==3.2.2
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2==2.9.7
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
redis==5.0.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4

